Switch to 'esm' for loading ES modules, use Babel transpiling for tests
Remove 'prepublish' dependency.
Fixes #48.
