+++ package.json
@@ -2,12 +2,11 @@
   "name": "autodux",
   "version": "5.0.0",
   "description": "Automate the Redux boilerplate.",
-  "browser": "dist/index.js",
-  "main": "dist/index.js",
-  "module": "dist/index-esm.js",
+  "main": "src/index.js",
+  "module": "src/core.js",
   "scripts": {
     "lint": "eslint . && echo 'Lint finished.'",
-    "test": "riteway src/test.js | tap-nirvana",
+    "test": "node -r @babel/register -r @babel/polyfill ./node_modules/.bin/riteway src/test.js | tap-nirvana",
     "test-coverage": "nyc npm test",
     "test-coverage-ci": "nyc --reporter=text-lcov npm test",
     "show-coverage-ci": "nyc report --reporter=text-lcov | coveralls",
@@ -15,7 +14,6 @@
     "show-coverage-html": "open ./coverage/index.html || echo \"Run 'npm run test-coverage' first.\"",
     "debug": "echo 'Open debugger in Chrome: \"chrome://inspect\".' && node --inspect-brk src/test.js",
     "watch": "watch 'clear && npm -s test' src",
-    "prepare": "npm run -s lint && npm -s test && prepublish",
     "update": "updtr"
   },
   "repository": {
@@ -34,6 +32,10 @@
   },
   "homepage": "https://github.com/ericelliott/autodux#readme",
   "devDependencies": {
+    "@babel/core": "7.3.3",
+    "@babel/polyfill": "7.2.5",
+    "@babel/preset-env": "7.3.1",
+    "@babel/register": "7.0.0",
     "coveralls": "3.0.3",
     "eslint": "5.14.1",
     "eslint-config-prettier": "4.0.0",
@@ -41,7 +43,6 @@
     "husky": "1.3.1",
     "lint-staged": "8.1.4",
     "nyc": "13.3.0",
-    "prepublish": "2.2.0",
     "prettier": "1.16.4",
     "riteway": "6.0.0",
     "tap-nirvana": "1.1.0",
@@ -49,6 +50,7 @@
     "watch": "1.0.2"
   },
   "dependencies": {
+    "esm": "3.2.5",
     "lodash.curry": "4.1.1"
   }
 }
