+++ src/errors.js
@@ -1,0 +1,1 @@
+export const SLICE_VALUE_ERROR = "Proper value of 'slice' is required!";
